+++ annimate_desktop/src/lib/api.ts
@@ -9,15 +9,15 @@ import {
   QueryNodesResult,
   QueryValidationResult,
 } from '@/lib/api-types';
-import { open as fileOpen, save } from '@tauri-apps/api/dialog';
-import { UnlistenFn, emit } from '@tauri-apps/api/event';
+import { Channel, invoke } from '@tauri-apps/api/core';
+import { emit } from '@tauri-apps/api/event';
 import { dirname } from '@tauri-apps/api/path';
-import { exit, relaunch } from '@tauri-apps/api/process';
-import { open as shellOpen } from '@tauri-apps/api/shell';
-import { invoke } from '@tauri-apps/api/tauri';
-import { appWindow } from '@tauri-apps/api/window';
+import { open as fileOpen, save } from '@tauri-apps/plugin-dialog';
+import { exit, relaunch } from '@tauri-apps/plugin-process';
+import { open as shellOpen } from '@tauri-apps/plugin-shell';
+import { check as checkForUpdate } from '@tauri-apps/plugin-updater';
 
-export { dirname, exit, fileOpen, relaunch, save, shellOpen };
+export { checkForUpdate, dirname, exit, fileOpen, relaunch, save, shellOpen };
 
 export const deleteCorpus = (params: { corpusName: string }): Promise<void> =>
   invoke('delete_corpus', params);
@@ -33,14 +33,25 @@ export const emitExportCancelRequestedEvent = (): Promise<void> =>
 export const emitImportCancelRequestedEvent = (): Promise<void> =>
   emit('import_cancel_requested');
 
-export const exportMatches = (params: {
-  corpusNames: string[];
-  aqlQuery: string;
-  queryLanguage: QueryLanguage;
-  exportColumns: ExportColumn[];
-  exportFormat: ExportFormat;
-  outputFile: string;
-}): Promise<void> => invoke('export_matches', params);
+export const exportMatches = async (
+  params: {
+    corpusNames: string[];
+    aqlQuery: string;
+    queryLanguage: QueryLanguage;
+    exportColumns: ExportColumn[];
+    exportFormat: ExportFormat;
+    outputFile: string;
+  },
+  handlers: {
+    onEvent?: (event: ExportStatusEvent) => void;
+  } = {},
+): Promise<void> => {
+  const eventChannel = new Channel<ExportStatusEvent>();
+  if (handlers.onEvent !== undefined) {
+    eventChannel.onmessage = handlers.onEvent;
+  }
+  await invoke('export_matches', { eventChannel, ...params });
+};
 
 export const addCorporaToSet = (params: {
   corpusSet: string;
@@ -67,8 +78,20 @@ export const getSegmentations = (params: {
   corpusNames: string[];
 }): Promise<string[]> => invoke('get_segmentations', params);
 
-export const importCorpora = (params: { paths: string[] }): Promise<string[]> =>
-  invoke('import_corpora', params);
+export const importCorpora = async (
+  params: {
+    paths: string[];
+  },
+  handlers: {
+    onEvent?: (event: ImportStatusEvent) => void;
+  } = {},
+): Promise<string[]> => {
+  const eventChannel = new Channel<ImportStatusEvent>();
+  if (handlers.onEvent !== undefined) {
+    eventChannel.onmessage = handlers.onEvent;
+  }
+  return await invoke('import_corpora', { eventChannel, ...params });
+};
 
 export const renameCorpusSet = (params: {
   corpusSet: string;
@@ -85,17 +108,3 @@ export const validateQuery = (params: {
   aqlQuery: string;
   queryLanguage: QueryLanguage;
 }): Promise<QueryValidationResult> => invoke('validate_query', params);
-
-export const subscribeToExportStatus = (
-  callback: (statusEvent: ExportStatusEvent) => void,
-): Promise<UnlistenFn> =>
-  appWindow.listen<ExportStatusEvent>('export_status', ({ payload }) =>
-    callback(payload),
-  );
-
-export const subscribeToImportStatus = (
-  callback: (statusEvent: ImportStatusEvent) => void,
-): Promise<UnlistenFn> =>
-  appWindow.listen<ImportStatusEvent>('import_status', ({ payload }) =>
-    callback(payload),
-  );